requirements.txt
streamlit==1.31.0
requests==2.31.0
openai==1.12.0
openpyxl==3.1.2
pyautogen==0.2.13
fuzzywuzzy==0.18.0
arrow==1.3.0